Work Changes
Need to test new grids, targeting, priority, strength

+++ McRave/BaseManager.h
@@ -9,10 +9,10 @@ using namespace std;
 class BaseTrackerClass
 {	
 	map <Unit, BaseInfo> myBases;
-	map <int, TilePosition> myOrderedBases;
+	map <double, TilePosition> myOrderedBases;
 public:
 	map <Unit, BaseInfo>& getMyBases() { return myBases; }
-	map <int, TilePosition>& getMyOrderedBases() { return myOrderedBases; }
+	map <double, TilePosition>& getMyOrderedBases() { return myOrderedBases; }
 
 	void update();
 	void updateAlliedBases();
